% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fullcalendar.R
\name{fullcalendar}
\alias{fullcalendar}
\title{Generate Calendar}
\usage{
fullcalendar(events = NULL, options = NULL, callbacks = NULL,
  width = NULL, height = NULL, elementId = NULL)
}
\arguments{
\item{events}{data.frame where each row is a event object, see \url{https://fullcalendar.io/docs/event_data/Event_Object/}}

\item{options}{list of options, see notes}

\item{callbacks}{list of JavaScript callback functions. Must wrap JavaScript with \code{htmlwidgets::JS}}

\item{width, height}{Must be a valid CSS unit (like \code{'100\%'},
\code{'400px'}, \code{'auto'}) or a number, which will be coerced to a
string and have \code{'px'} appended.}

\item{elementId}{Use an explicit element ID for the widget (rather than an automatically generated one). Useful if you have other JavaScript that needs to explicitly discover and interact with a specific widget instance.}
}
\description{
Generate calendar
}
\note{
Documentation for options and callbacks is available at \url{https://fullcalendar.io/docs/}
}
\examples{
\dontrun{
# blank calender of current month
fullcalendar()

# with events
fullcalendar(events = data.frame(title = "Event", date = Sys.Date()))

# with options and callbacks
fullcalendar(options = list(header = list(left = "", center = "", right = "prev,next")),
             callbacks = list(dayClick = DT::JS(
             "function(date, jsEvent, view) {
               alert(date.format());
             }"
             )))
}
}
